front-components/requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask-Mobility==0.1.1
Flask-WTF==0.14.2
httplib2==0.12.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
PyJWT==1.7.1
PySocks==1.6.8
pytz==2018.9
requests==2.21.0
six==1.12.0
twilio==6.23.1
urllib3==1.24.1
Werkzeug==0.14.1
WTForms==2.2.1
